+++ src/main/webapp/resources/js/views/SearchPanel.js
@@ -42,6 +42,8 @@ Ext.define('BagDatabase.views.SearchPanel', {
             fieldLabel: 'Full Text Search',
             itemId: 'searchTerms',
             name: 'searchTerms',
+            stateful: true,
+            stateId: 'searchTerms',
             flex: 1,
             margin: 5,
             enableKeyEvents: true,
@@ -66,6 +68,7 @@ Ext.define('BagDatabase.views.SearchPanel', {
             xtype: 'button',
             itemId: 'searchButton',
             text: 'Search',
+            iconCls: 'magnifier-icon',
             margin: 5,
             handler: function(button) {
                 var vp = button.up('viewport');
@@ -74,16 +77,23 @@ Ext.define('BagDatabase.views.SearchPanel', {
                 var fields = vp.down('#searchFields').getValue();
                 store.filterBags(terms, fields);
             }
-        }, { xtype: 'tbseparator' }, {
-            xtype: 'navigationButton',
-            isAdmin: isAdmin,
-            margin: 5
         }]
     }, {
         xtype: 'checkboxgroup',
         itemId: 'searchFields',
+        stateful: true,
+        stateId: 'gridSearchFields',
         anchor: '100%',
         columns: 7,
+        // The checkboxgroup widget doesn't actually save the states of its
+        // checkboxes by default, so we have to manually do that.
+        stateEvents: ['change'],
+        applyState: function(state) {
+            this.setValue(state);
+        },
+        getState: function() {
+            return this.getValue();
+        },
         getErrors: function(value) {
             var errors = [];
             var tnCheckbox = this.down('#topicNameCheckbox');